rename from DotNetTransfer to ContainerRegistryTransfer
