+++ samples/dotnetcore/image-transfer/ContainerRegistryTransfer/ContainerRegistryTransfer.csproj
