+++ samples/dotnetcore/image-transfer/ContainerRegistryTransfer/Helpers/IdentityHelper.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace DotNetTransfer.Helpers
+namespace ContainerRegistryTransfer.Helpers
 {
     public static class IdentityHelper
     {
