+++ samples/dotnetcore/image-transfer/ContainerRegistryTransfer/Helpers/KeyVaultHelper.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Configuration;
 using System;
 using System.Collections.Generic;
 
-namespace DotNetTransfer.Helpers
+namespace ContainerRegistryTransfer.Helpers
 {
     public class KeyVaultHelper
     {
