+++ samples/dotnetcore/image-transfer/ContainerRegistryTransfer/Options.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace DotNetTransfer
+namespace ContainerRegistryTransfer
 {
     public class Options
     {
