+++ samples/dotnetcore/image-transfer/ContainerRegistryTransfer/appsettings.json
